It boots! (again!)
Got the application to booting again. It just crashes when you click on
the record button and try to start the media recorder.

Notes:
-Make sure google_play_services_lib is seen as a library in the
properties of your project.
-Also remove it from the projects that are in the Android build. If it
is in there it will go looking for a non existent .apk file for the
lib.

Just working on getting the recording working again. it is still in 3gg
format unfortunately. Still need to look into recording/converting into
ogg. Sorry about this, just been caught up in trying to get it to run
at all first.