Map improvements
Changed which buttons appear on the map. Have hopefully made the map
centre on the user on open. Few other improvements to the map. Got badly
distracted by Rust so there is a little less here than hoped. Also
recording icon (temp of course).

Main thing to be done after getting the map and recording working is to
bundle up the data to send to the server