Complete re-write
I made the classic mistake of jumping into something without having an
idea of the structure I wanted to create. After having a slight meltdown
on tuesday/wednesday about this I started a fresh with a new plan.

What we have now is an app that has a flow through its basic systems and
is pretty much half feature complete.

Still to do
-Camera integration (difficulty: hard)
-Upload page (difficulty:low)
-Put data into JSON (difficulty:unknown. Not looked into it yet)

And testing for location data grabbing and audio recording.