UploadActivity work
built the upload activity out a fair bit. Sorted out its action bar,
created the layout to show all the data.

Still to do in the upload page
-Handle exit and upload buttons
-exit has confirm dialog then drops to start
-upload will package up the JSON and store it with the audio file. This
will be extendible to upload OTA
-openSettings needs to be dealt with
-Image gallery needs to be handled.

Apart from this the 2 main things that need to be dealt with are
-picture capture
-background recording

Need to do some testing but it looks like it is all dependent on getting
all the parts in first