failed refactor
Tried to move the app to Android Studio to be able to properly make a
Navigation Drawer. After 2.5 hours of hassle this experiment will be
left. What should be a relatively small task (importing a library) has
caused so many issues. Got it imported to then get stuck on XML issues.

Eclipse has miles better error reporting and help functions. It may not
be as pretty, but at least it runs faster than a snails pace!

Will continue work on improving location data gathering tonight.