getMediaOutputFile issues
Pretty much all the camera impl is done (layout, calling it etc) but I
am having a problem storing the images (ah, as always. Storage issues).

Going to restructure the recordingActivity directory creation stuff but
have worked for about 12 hours on this today so bedtime for now. To be
continued.